+++ scripts/calibnet_health_check.sh
@@ -32,17 +32,16 @@ echo "Fetching params"
 $FOREST_CLI_PATH fetch-params --keys
 echo "Downloading zstd compressed snapshot"
 $FOREST_CLI_PATH --chain calibnet snapshot fetch --aria2 --provider filecoin --compressed -s "$SNAPSHOT_DIRECTORY"
-# Temporarily disabled because there might be a problem with the filecoin snapshots
-#echo "Importing zstd compressed snapshot"
-#$FOREST_PATH --chain calibnet --encrypt-keystore false --halt-after-import --height=-200 --import-snapshot "$SNAPSHOT_DIRECTORY"/*.zst
-#echo "Cleaning up database"
-#$FOREST_CLI_PATH --chain calibnet db clean --force
-#echo "Importing snapshot from url"
-#$FOREST_PATH --chain calibnet --encrypt-keystore false --halt-after-import --height=-200 --import-snapshot https://snapshots.calibrationnet.filops.net/minimal/latest
-#echo "Cleaning up database"
-#$FOREST_CLI_PATH --chain calibnet db clean --force
-#echo "Importing zstd compressed snapshot from url"
-#$FOREST_PATH --chain calibnet --encrypt-keystore false --halt-after-import --height=-200 --import-snapshot https://snapshots.calibrationnet.filops.net/minimal/latest.zst
+echo "Importing zstd compressed snapshot"
+$FOREST_PATH --chain calibnet --encrypt-keystore false --halt-after-import --height=-200 --import-snapshot "$SNAPSHOT_DIRECTORY"/*.zst
+echo "Cleaning up database"
+$FOREST_CLI_PATH --chain calibnet db clean --force
+echo "Importing snapshot from url"
+$FOREST_PATH --chain calibnet --encrypt-keystore false --halt-after-import --height=-200 --import-snapshot https://snapshots.calibrationnet.filops.net/minimal/latest
+echo "Cleaning up database"
+$FOREST_CLI_PATH --chain calibnet db clean --force
+echo "Importing zstd compressed snapshot from url"
+$FOREST_PATH --chain calibnet --encrypt-keystore false --halt-after-import --height=-200 --import-snapshot https://snapshots.calibrationnet.filops.net/minimal/latest.zst
 echo "Cleaning up database"
 $FOREST_CLI_PATH --chain calibnet db clean --force